recipeservicetests make use of logger and move up dispose method
